backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0